Refactor imports for consistency and clarity across multiple files; add BFGPublisher to the module exports
