+++ CityOfBinds/src/content_managers/templates/utils/command_factory.py
@@ -1,4 +1,4 @@
-from .....utils.templates.templates import StringTemplate, Pool
+from .....utils import StringTemplate, Pool
 
 
 class _CommandFactory(StringTemplate):
