+++ CityOfBinds/src/content_managers/templates/utils/commands_template.py
@@ -1,7 +1,7 @@
 from .command_factory import _CommandFactory
 from ....game.utils.powers import _Power
 from ....game.utils.slash_commands import _CommandGroup, CommandGroupConstants
-from .....utils.templates import ListTemplate, Pool
+from .....utils import ListTemplate, Pool
 
 
 class _CommandsTemplate(_CommandGroup, ListTemplate):
