+++ CityOfBinds/src/content_publisher/__init__.py
@@ -1,0 +1,5 @@
+from .bfg_publisher import BFGPublisher
+
+__all__ = [
+    "BFGPublisher",
+]
