--- CityOfBinds/utils/templates/__init__.py
@@ -1,4 +1,0 @@
-from .pool import Pool
-from .templates import StringTemplate, ListTemplate
-
-__all__ = ["Pool", "StringTemplate", "ListTemplate"]
